feat: when electron preload build, reload renderer

+++ src/main.ts
@@ -1,6 +1,8 @@
+import type { ChildProcess } from 'node:child_process';
 import type { Options as TsupOptions } from 'tsup';
+import type { ViteDevServer } from 'vite';
 import type { InnerOptions, PluginOptions } from './types';
-import { ChildProcess, spawn } from 'node:child_process';
+import { spawn } from 'node:child_process';
 import electron from 'electron';
 import { lightGreen } from 'kolorist';
 import { build as tsupBuild } from 'tsup';
@@ -8,7 +10,7 @@ import { createLogger } from './logger';
 
 const logger = createLogger();
 
-function getBuildOptions(options: PluginOptions, innerOpts: InnerOptions): TsupOptions[] {
+function getBuildOptions(options: PluginOptions, innerOpts: InnerOptions) {
   const env = {
     APP_DEV_SERVER_URL: innerOpts?.serverUrl,
   };
@@ -21,13 +23,16 @@ function getBuildOptions(options: PluginOptions, innerOpts: InnerOptions): TsupO
 
   return ['main', 'preload']
     .filter(s => options[s] && options[s].entry)
-    .map(s => options[s])
+    .map(s => {
+      options[s].__NAME__ = s;
+      return options[s];
+    })
     .map(cfg => {
       return {
         ...cfg,
         env,
         silent: true,
-      } as TsupOptions;
+      } as TsupOptions & { __NAME__: string };
     });
 }
 
@@ -45,13 +50,17 @@ function runMainProcess(options: PluginOptions, innerOpts: InnerOptions) {
   }).on('exit', exitMainProcess);
 }
 
-export async function runServe(options: PluginOptions, innerOpts: InnerOptions) {
-  let electronProcess: ChildProcess;
+export async function runServe(
+  options: PluginOptions,
+  innerOpts: InnerOptions,
+  server: ViteDevServer,
+) {
+  let mainProcess: ChildProcess;
 
   const killProcess = () => {
-    if (electronProcess) {
-      electronProcess.off('exit', exitMainProcess);
-      electronProcess.kill();
+    if (mainProcess) {
+      mainProcess.off('exit', exitMainProcess);
+      mainProcess.kill();
     }
   };
 
@@ -59,40 +68,42 @@ export async function runServe(options: PluginOptions, innerOpts: InnerOptions) 
     killProcess();
   });
 
-  logger.info(`electron server start`);
-
   const buildOptions = getBuildOptions(options, innerOpts);
 
   for (let i = 0; i < buildOptions.length; i++) {
     let isFirstBuild = true;
     const tsOpts = buildOptions[i];
-    const { onSuccess: _onSuccess, watch, ...tsupOptions } = tsOpts;
+    const { __NAME__: name, onSuccess: _onSuccess, watch, ...tsupOptions } = tsOpts;
 
-    logger.info(`electron ${tsOpts.name} process build`);
+    logger.info(`${name} build`);
 
     const onSuccess: TsupOptions['onSuccess'] = async () => {
-      logger.info(`electron ${tsOpts.name} process build success`);
-
       if (typeof _onSuccess === 'function') {
         await _onSuccess();
       }
 
-      logger.success('rebuild succeeded!');
-
       if (isFirstBuild) {
+        logger.info(`${name} build succeeded`);
         isFirstBuild = false;
         return;
       }
 
-      killProcess();
+      logger.success(`${name} rebuild succeeded!`);
 
-      electronProcess = runMainProcess(options, innerOpts);
+      if (name === 'main') {
+        killProcess();
+        mainProcess = runMainProcess(options, innerOpts);
+      } else {
+        server.ws.send({
+          type: 'full-reload',
+        });
+      }
     };
 
     await tsupBuild({ onSuccess, watch: true, ...tsupOptions });
   }
 
-  electronProcess = runMainProcess(options, innerOpts);
+  mainProcess = runMainProcess(options, innerOpts);
 
   return {
     kill: killProcess,
